requirements.txt
djangorestframework
markdown
django-filter
recombee-api-client
djangorestframework-simplejwt
Pillow
django-cors-headers
torch>=1.12.0
torchvision>=0.13.0
sentence-transformers>=2.2.2
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.62.0
faiss-cpu>=1.7.4
pytest